testcode/tcode/testapp/requirements.txt
appier==1.18.25
appnope==0.1.0
asgiref==3.2.3
astroid==2.3.2
asyncio-redis==0.15.1
backcall==0.1.0
BareNecessities==0.2.8
beautifulsoup4==4.8.0
boto3==1.9.220
botocore==1.12.220
bs4==0.0.1
cachetools==3.1.0
cairocffi==1.1.0
CairoSVG==2.4.2
canvas==0.2
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
class-registry==2.1.2
Click==7.0
configparser==4.0.2
css-html-js-minify==2.5.5
cssselect==1.1.0
cssselect2==0.2.2
cssutils==1.0.2
cycler==0.10.0
databases==0.2.6
decorator==4.4.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.1.15
django-allauth==0.39.1
django-auth-ldap==2.0.0
django-ckeditor==5.7.1
django-core==1.4.1
django-cors-headers==3.1.0
django-crispy-forms==1.8.0
django-debug-toolbar==2.0
django-extensions==2.2.5
django-filter==1.1.0
django-filters==0.1.0
django-js-asset==1.2.2
django-sendgrid-v5==0.8.0
django-shell-plus==1.1.7
django-storages==1.7.1
django-utils==0.0.2
django-weasyprint==0.5.4
djangorestframework==3.11.0
djangorestframework-filters==0.11.1
docopt==0.6.2
docutils==0.15.2
Flask==1.0.3
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.4.0
future==0.17.1
google-api==0.1.12
google-api-core==1.11.1
google-api-python-client==1.7.9
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud==0.34.0
google-cloud-core==1.0.1
google-cloud-datastore==1.8.0
google-cloud-monitoring==0.31.1
google-cloud-pubsub==0.39.1
google-cloud-storage==1.18.0
google-cloud-translate==1.5.0
google-resumable-media==0.3.2
google-translate==0.1
googleapis-common-protos==1.5.8
grpc-google-iam-v1==0.11.4
grpcio==1.19.0
gunicorn==20.0.0
html5lib==1.0.1
htmlmin==0.1.12
httplib2==0.13.0
idna==2.8
ipython==7.8.0
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.1
jmespath==0.9.4
kiwisolver==1.1.0
lazy-object-proxy==1.4.2
ld==0.5.0
lxml==4.4.1
Mail==2.1.0
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
mysql==0.0.2
mysqlclient==1.4.6
numpy==1.16.4
oauthlib==3.0.2
pandas==0.24.2
pangocffi==0.5.0
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
post==2019.4.13
posts==0.2.2
premailer==3.6.1
prompt-toolkit==2.0.10
protobuf==3.7.0
psycopg2-binary==2.8.3
ptyprocess==0.6.0
pyango-view==0.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.19
Pygments==2.4.2
pylint==2.4.3
PyMySQL==0.9.3
pyparsing==2.4.4
pyPdf==1.13
PyPDF2==1.26.0
Pyphen==0.9.5
python-dateutil==2.8.0
python-http-client==3.2.1
python-ldap==3.2.0
python3-openid==3.1.0
pythonping==1.0.3
pytz==2019.3
PyYAML==5.1.2
query==0.1.4
reportlab==3.5.32
reportlib==3.4.0
requests==2.21.0
requests-oauthlib==1.2.0
resolver==0.2.1
reverse==0.1.0
rsa==4.0
s3transfer==0.2.1
semver==2.7.2
sendgrid==6.1.0
six==1.12.0
socker==2.0.1
soupsieve==1.9.2
SQLAlchemy==1.3.4
sqlparse==0.3.0
tinycss2==1.0.2
traitlets==4.3.3
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.24.1
wcwidth==0.1.7
WeasyPrint==49
webencodings==0.5.1
websockets==7.0
Werkzeug==0.15.4
whatap-python==0.1.73
whitenoise==4.1.4
wrapt==1.11.2
xhtml2pdf==0.2.3
